Set perm bits to 0777 (rwx) for all (ugo)
Explicitly set permission bits to 0777 to make SCATR binaries executable by default as until now, 0644 was used by default, i.e., executable bit was not set for all (ugo). Setting 0777 so that it is readable, modifiable, and also executable for all (user, group, others). If any user wishes to modify the permissions, they can do it independently. But by default, we should ship the binaries with 0777.

Signed-off-by: subham sarkar <80461438+subham-deepsource@users.noreply.github.com>